Added config for gpu/cpu, and numeric/analytic psi

+++ src/dirac.rs
@@ -1,22 +1,38 @@
 //! Scratchpad for Dirac-equation implementation.
 //! Intent: See if this clarifies how to handle spin.
 //!
-//! Relation of E and p for any particle: E^2 - phat^2 c^c = m^2 c^4
+//! Relation of E and p̂ for any particle: E^2 - p̂^2 c^c = m^2 c^4
 //!
 //! //! α_1^2 = α_2^2 = α_3^2 = β^2 = 1
 //! α and β are 4x4 matrices. α = [0,σ, σ, 0]  β = [1, 0, 0, -1]
 //!
-//! i hbar * dψ/dt = H_dirac ψ
-//! H_dirac = cα (p_hat + e/C A) + βmc^2 + V(r)   V(r) ~= -e^2/r = (-e) * ϕ(r)
+//! iħ * dψ/dt = H_dirac ψ
+//! H_dirac = cα (p̂ + e/C A) + βmc^2 + V(r)   V(r) ~= -e^2/r = (-e) * ϕ(r)
 //! ϕ is the scaler potential, eg like V.
 //!
-//! ψ = [X, ϕ] X: Pauli.
+//! ψ = [X, ϕ] X: Pauli.  [energy eigenstate. X and ϕ are 2-component spinors]
 //! HX = EX.
 //!
-//! H = phat^2/2m + V(r) [H0] - phat^4/(8m^3c^2) [(dH)_relativistic] +
-//! 1/(2m^2c^2)(1/r)(dV/dr)S L [spin orbit coupling]
+//! p̂ = -iħ∇
 //!
-
+//! H = p̂^2/2m + V(r) [H0: Schrodinger eq] - p̂^4/(8m^3c^2) [(dH)_relativistic] +
+//! 1/(2m^2c^2)(1/r)(dV/dr)S L [spin orbit coupling; S is spin angular momentum; L is orbital
+//! angular momentum] + ħ^2 / (8m^2c^2) ∇^2 V [dH Darwin]
+//!
+//! Perterbation theory...
+//!
+//! https://arxiv.org/ftp/arxiv/papers/1709/1709.04338.pdf
+//!
+//!
+//! "The density of electronic charge, which is
+//! proportional to the square of Schroedinger's scalar amplitude function for the hydrogen atom, or
+//! precisely a product of a function and its complex conjugate as ψ* ψ to take account of the
+//! complex character, hence became proportional to a scalar product of Dirac's vectorial amplitude
+//! function for each state with its complex conjugate vector."
+//!
+//! (For Darwin term) Vtilde(r) = V(r) + 1/10 * hbar^2/(m^2 c^2) * nalba^2 V
+//!
+//! Dirac psi amplitude: 4 components: 2 large; 2 small (vector)
 
 use na::{Matrix2, Matrix4};
 use nalgebra as na;
@@ -29,12 +45,12 @@ const C0: Cplx = Cplx::new_zero();
 
 #[rustfmt::skip]
 fn a() {
-    let gamma: Matrix2<f32> = Matrix2::new(
+    let gamma: Matrix2<f64> = Matrix2::new(
         0.0, 0.0,
         0.0, 0.0,
     );
 
-    let beta: Matrix4<f32> = Matrix4::new(
+    let beta: Matrix4<f64> = Matrix4::new(
         0.0, 0.0,0.0, 0.0,
         0.0, 0.0,0.0, 0.0,
         0.0, 0.0,0.0, 0.0,
@@ -42,14 +58,14 @@ fn a() {
     );
 
     // [Gamma matrices](https://en.wikipedia.org/wiki/Gamma_matrices)
-    let gamma0: Matrix4<f32> = Matrix4::new(
+    let gamma0: Matrix4<f64> = Matrix4::new(
         1., 0., 0., 0.,
         0., 1., 0., 0.,
         0., 0., -1., 0.,
         0., 0., 0., -1.
     );
 
-    let gamma1: Matrix4<f32> = Matrix4::new(
+    let gamma1: Matrix4<f64> = Matrix4::new(
         0., 0., 0., 1.,
         0., 0., 1., 0.,
         0., -1., 0., 0.,
@@ -63,7 +79,7 @@ fn a() {
         -IM, C0, C0, C0
     );
 
-    let gamma3: Matrix4<f32> = Matrix4::new(
+    let gamma3: Matrix4<f64> = Matrix4::new(
         1., 0., 1., 0.,
         0., 0., 0., -1.,
         -1., 0., 0., 0.,
@@ -71,7 +87,7 @@ fn a() {
     );
 
     // The identity matrix.
-    let gamma4: Matrix4<f32> = Matrix4::new(
+    let gamma4: Matrix4<f64> = Matrix4::new(
         1., 0., 0., 0.,
         0., 1., 0., 0.,
         0., 0., 1., 0.,
@@ -79,7 +95,7 @@ fn a() {
     );
 
     // γ5 is not a proper member of the gamma group.
-    let gamma5: Matrix4<f32> = Matrix4::new(
+    let gamma5: Matrix4<f64> = Matrix4::new(
         0., 0., 1., 0.,
         0., 0., 0., 1.,
         1., 0., 0., 0.,
@@ -87,24 +103,23 @@ fn a() {
     );
 
     // [Pauli matrices](https://en.wikipedia.org/wiki/Pauli_matrices)
-    let sigma1: Matrix2<f32> = Matrix2::new(
+    let sigma1: Matrix2<f64> = Matrix2::new(
         0.0, 1.,
         1., 0.0,
     );
 
-    // [Pauli matrices](https://en.wikipedia.org/wiki/Pauli_matrices)
     let sigma2: Matrix2<Cplx> = Matrix2::new(
         C0, -IM,
         IM, C0,
     );
 
-    // [Pauli matrices](https://en.wikipedia.org/wiki/Pauli_matrices)
-    let sigma3: Matrix2<f32> = Matrix2::new(
+    let sigma3: Matrix2<f64> = Matrix2::new(
         1., 0.,
         0., -1.,
     );
 
-    let alpha1: Matrix4<f32> = Matrix4::new(
+    // Note: The `m11` syntax works for `Matrix4<f64>`, but not <Cplx>.
+    let alpha1: Matrix4<f64> = Matrix4::new(
         0., 0., sigma1[(1, 1)], sigma1[(1, 2)],
         0., 0., sigma1[(2, 1)], sigma1[(2, 2)],
         sigma1[(1, 1)], sigma1[(1, 2)], 0., 0.,
@@ -118,15 +133,14 @@ fn a() {
         sigma2[(2, 1)], sigma2[(2, 2)], C0, C0
     );
 
-    let alpha3: Matrix4<f32> = Matrix4::new(
+    let alpha3: Matrix4<f64> = Matrix4::new(
         0., 0., sigma3[(1, 1)], sigma3[(1, 2)],
         0., 0.,sigma3[(2, 1)], sigma3[(2, 2)],
         sigma3[(1, 1)], sigma3[(1, 2)], 0., 0.,
         sigma3[(2, 1)], sigma3[(2, 2)], 0., 0.
     );
     
-    // etc for alpha 2 and 3.
-    let beta: Matrix4<f32> = Matrix4::new(
+    let beta: Matrix4<f64> = Matrix4::new(
         1., 0., 0., 0.,
         0., 1., 0., 0.,
         0., 0., -1., 0.,
